pulling common worker code out of the exposure package
