+++ python/backend/backend/exposure/config.py
@@ -10,22 +10,6 @@ DEFAULT_REFUTATION_TESTS = ["random_common_cause", "placebo_treatment_refuter"]
 SENSITIVITY_REFUTERS = ["add_unobserved_common_cause"]
 
 
-def get_redis_url():
-    redis_url = os.environ.get("REDIS_URL")
-    if redis_url:
-        return redis_url
-    else:
-        raise Exception("REDIS_URL configuration missing in environment")
-
-
-def get_storage_url():
-    storage_location = os.environ.get("STORAGE")
-    if storage_location:
-        return storage_location
-    else:
-        raise Exception("STORAGE configuration missing in environment")
-
-
 def get_refuters():
     include_sensitivity_refuters = (
         os.environ.get("INCLUDE_SENSITIVITY_REFUTERS", "false").strip().lower()
