+++ python/backend/backend/exposure/worker/significance_test_task.py
@@ -13,12 +13,12 @@ from backend.exposure.inference.significance_test import (
     compute_null_effect,
     get_propensity_scores,
 )
-from backend.exposure.io.db import get_db_client
 from backend.exposure.model.significance_test_models import (
     ComputeNullEffectSpec,
     PropensityScoreSpec,
 )
 from backend.exposure.worker.worker import exposure_worker
+from backend.worker_commons.io.db import get_db_client
 
 
 @exposure_worker.task
